set fw to empty string when it does not exist

+++ update-gh-pages.sh
@@ -114,7 +114,12 @@ push_changes()
 
   cd $HOME/gh-pages-$TT_BUILD/download
 
-  cp -Rf $image/$fw .
+  if [ -f $image/$fw ]; then
+    cp -Rf $image/$fw .
+  else
+    echo "$fw not found!"
+    fw=""
+  fi
   # cd $HOME/gh-pages-$TT_BUILD/download
 
   #add, commit and push files
